+++ src/Accelerators.cpp
@@ -123,6 +123,7 @@ static ACCEL gBuiltInAccelerators[] = {
     {FSHIFT | FVIRTKEY, VK_F11, CmdTogglePresentationMode},
     {FSHIFT | FCONTROL | FVIRTKEY, 'L', CmdToggleFullscreen},
     {FVIRTKEY, VK_F11, CmdToggleFullscreen},
+    {FSHIFT | FCONTROL | FVIRTKEY, 'H', CmdToggleAutomaticallyScroll},
     {FVIRTKEY, VK_F12, CmdToggleBookmarks},
     {FSHIFT | FVIRTKEY, VK_F12, CmdCommandPaletteTOC},
     {FSHIFT | FCONTROL | FVIRTKEY, VK_SUBTRACT, CmdRotateLeft},